Consolidating zoomadd and zoom functionality into add_zoom
Changed how the command tracks a bit, if the command fails to update the
counts or the times for any reason, it will now to rollback to the
original count and times.

Also added in a rolling window tracker to always state the number of
zooms within the past 7 days.